Isolate JSON and HTTP modules
This introduced the concept the client module to execute requests. Our
internal request object becomes a description of an HTTP request, while
the client uses a driver (for now only Net::HTTP) to execute it.

We now have the following isolation:

* Client isolates HTTP interactions
* Drivers::NetHttp isolates Net::HTTP dependency
* JsonClient isolates JSON-specific request and response conversions

By introducing a shortcut method in the client to build and execute a
request, we no longer the separate Media module that previously only
built requests.