Prevent empty schedule
Poms returns a 404 when there is nothing currently on tv for the /now and /next channel endpoints. This caused problems in out apps, so we now use the schedule endpoint to either return the most recent thing on tv or the next thing on. This should prevent the endpoint from returning nothing (or a 404).